Move Paginator component to molecule and delete uncessary libraries
Fixes #1398
